moved around insert operator code
