+++ go/vt/sqlparser/ast_funcs.go
@@ -2533,7 +2533,7 @@ func (v *visitor) visitAllSelects(in SelectStatement, f func(p *Select, idx int)
 	panic("switch should be exhaustive")
 }
 
-func IsNonLiteral(updExprs UpdateExprs) bool {
+func (updExprs UpdateExprs) IsNonLiteral() bool {
 	for _, updateExpr := range updExprs {
 		switch updateExpr.Expr.(type) {
 		case *Argument, *NullVal, BoolVal, *Literal:
@@ -2542,4 +2542,14 @@ func IsNonLiteral(updExprs UpdateExprs) bool {
 		}
 	}
 	return false
+}
+
+// IsRestrict returns true if the reference action is of restrict type.
+func (ra ReferenceAction) IsRestrict() bool {
+	switch ra {
+	case Restrict, NoAction, DefaultAction:
+		return true
+	default:
+		return false
+	}
 }
