+++ go/vt/vtgate/planbuilder/operators/insert.go
@@ -17,9 +17,15 @@ limitations under the License.
 package operators
 
 import (
+	"strconv"
+
+	vschemapb "vitess.io/vitess/go/vt/proto/vschema"
 	"vitess.io/vitess/go/vt/sqlparser"
+	"vitess.io/vitess/go/vt/vterrors"
+	"vitess.io/vitess/go/vt/vtgate/engine"
 	"vitess.io/vitess/go/vt/vtgate/evalengine"
 	"vitess.io/vitess/go/vt/vtgate/planbuilder/operators/ops"
+	"vitess.io/vitess/go/vt/vtgate/planbuilder/plancontext"
 	"vitess.io/vitess/go/vt/vtgate/vindexes"
 )
 
@@ -120,4 +126,308 @@ func (i *Insert) TablesUsed() []string {
 
 func (i *Insert) Statement() sqlparser.Statement {
 	return i.AST
+}
+
+func createOperatorFromInsert(ctx *plancontext.PlanningContext, ins *sqlparser.Insert) (ops.Operator, error) {
+	tableInfo, qt, err := createQueryTableForDML(ctx, ins.Table, nil)
+	if err != nil {
+		return nil, err
+	}
+
+	vindexTable, routing, err := buildVindexTableForDML(ctx, tableInfo, qt, "insert")
+	if err != nil {
+		return nil, err
+	}
+
+	if _, target := routing.(*TargetedRouting); target {
+		return nil, vterrors.VT12001("INSERT with a target destination")
+	}
+
+	insOp := &Insert{
+		VTable: vindexTable,
+		AST:    ins,
+	}
+	route := &Route{
+		Source:  insOp,
+		Routing: routing,
+	}
+
+	// Find the foreign key mode and store the ParentFKs that we need to verify.
+	ksMode, err := ctx.VSchema.ForeignKeyMode(vindexTable.Keyspace.Name)
+	if err != nil {
+		return nil, err
+	}
+	if ksMode == vschemapb.Keyspace_FK_MANAGED {
+		parentFKs := vindexTable.ParentFKsNeedsHandling(ctx.VerifyAllFKs, ctx.ParentFKToIgnore)
+		if len(parentFKs) > 0 {
+			return nil, vterrors.VT12002()
+		}
+	}
+
+	// Table column list is nil then add all the columns
+	// If the column list is empty then add only the auto-inc column and
+	// this happens on calling modifyForAutoinc
+	if ins.Columns == nil && valuesProvided(ins.Rows) {
+		if vindexTable.ColumnListAuthoritative {
+			ins = populateInsertColumnlist(ins, vindexTable)
+		} else {
+			return nil, vterrors.VT09004()
+		}
+	}
+
+	// modify column list or values for autoincrement column.
+	autoIncGen, err := modifyForAutoinc(ins, vindexTable)
+	if err != nil {
+		return nil, err
+	}
+	insOp.AutoIncrement = autoIncGen
+
+	// set insert ignore.
+	insOp.Ignore = bool(ins.Ignore) || ins.OnDup != nil
+
+	insOp.ColVindexes = getColVindexes(insOp)
+	switch rows := ins.Rows.(type) {
+	case sqlparser.Values:
+		route.Source, err = insertRowsPlan(insOp, ins, rows)
+		if err != nil {
+			return nil, err
+		}
+	case sqlparser.SelectStatement:
+		route.Source, err = insertSelectPlan(ctx, insOp, ins, rows)
+		if err != nil {
+			return nil, err
+		}
+	}
+	return route, nil
+}
+
+func insertSelectPlan(ctx *plancontext.PlanningContext, insOp *Insert, ins *sqlparser.Insert, sel sqlparser.SelectStatement) (*Insert, error) {
+	if columnMismatch(insOp.AutoIncrement, ins, sel) {
+		return nil, vterrors.VT03006()
+	}
+
+	selOp, err := PlanQuery(ctx, sel)
+	if err != nil {
+		return nil, err
+	}
+
+	// select plan will be taken as input to insert rows into the table.
+	insOp.Input = selOp
+
+	// When the table you are steaming data from and table you are inserting from are same.
+	// Then due to locking of the index range on the table we might not be able to insert into the table.
+	// Therefore, instead of streaming, this flag will ensure the records are first read and then inserted.
+	insertTbl := insOp.TablesUsed()[0]
+	selTables := TablesUsed(selOp)
+	for _, tbl := range selTables {
+		if insertTbl == tbl {
+			insOp.ForceNonStreaming = true
+			break
+		}
+	}
+
+	if len(insOp.ColVindexes) == 0 {
+		return insOp, nil
+	}
+
+	colVindexes := insOp.ColVindexes
+	vv := make([][]int, len(colVindexes))
+	for idx, colVindex := range colVindexes {
+		for _, col := range colVindex.Columns {
+			err := checkAndErrIfVindexChanging(sqlparser.UpdateExprs(ins.OnDup), col)
+			if err != nil {
+				return nil, err
+			}
+
+			colNum := findColumn(ins, col)
+			// sharding column values should be provided in the insert.
+			if colNum == -1 && idx == 0 {
+				return nil, vterrors.VT09003(col)
+			}
+			vv[idx] = append(vv[idx], colNum)
+		}
+	}
+	insOp.VindexValueOffset = vv
+	return insOp, nil
+}
+
+func columnMismatch(gen *Generate, ins *sqlparser.Insert, sel sqlparser.SelectStatement) bool {
+	origColCount := len(ins.Columns)
+	if gen != nil && gen.added {
+		// One column got added to the insert query ast for auto increment column.
+		// adjusting it here for comparison.
+		origColCount--
+	}
+	if origColCount < sel.GetColumnCount() {
+		return true
+	}
+	if origColCount > sel.GetColumnCount() {
+		sel := sqlparser.GetFirstSelect(sel)
+		var hasStarExpr bool
+		for _, sExpr := range sel.SelectExprs {
+			if _, hasStarExpr = sExpr.(*sqlparser.StarExpr); hasStarExpr {
+				break
+			}
+		}
+		if !hasStarExpr {
+			return true
+		}
+	}
+	return false
+}
+
+func insertRowsPlan(insOp *Insert, ins *sqlparser.Insert, rows sqlparser.Values) (*Insert, error) {
+	for _, row := range rows {
+		if len(ins.Columns) != len(row) {
+			return nil, vterrors.VT03006()
+		}
+	}
+
+	if len(insOp.ColVindexes) == 0 {
+		return insOp, nil
+	}
+
+	colVindexes := insOp.ColVindexes
+	routeValues := make([][][]evalengine.Expr, len(colVindexes))
+	for vIdx, colVindex := range colVindexes {
+		routeValues[vIdx] = make([][]evalengine.Expr, len(colVindex.Columns))
+		for colIdx, col := range colVindex.Columns {
+			err := checkAndErrIfVindexChanging(sqlparser.UpdateExprs(ins.OnDup), col)
+			if err != nil {
+				return nil, err
+			}
+			routeValues[vIdx][colIdx] = make([]evalengine.Expr, len(rows))
+			colNum, _ := findOrAddColumn(ins, col)
+			for rowNum, row := range rows {
+				innerpv, err := evalengine.Translate(row[colNum], nil)
+				if err != nil {
+					return nil, err
+				}
+				routeValues[vIdx][colIdx][rowNum] = innerpv
+			}
+		}
+	}
+	// here we are replacing the row value with the argument.
+	for _, colVindex := range colVindexes {
+		for _, col := range colVindex.Columns {
+			colNum, _ := findOrAddColumn(ins, col)
+			for rowNum, row := range rows {
+				name := engine.InsertVarName(col, rowNum)
+				row[colNum] = sqlparser.NewArgument(name)
+			}
+		}
+	}
+	insOp.VindexValues = routeValues
+	return insOp, nil
+}
+
+func valuesProvided(rows sqlparser.InsertRows) bool {
+	switch values := rows.(type) {
+	case sqlparser.Values:
+		return len(values) >= 0 && len(values[0]) > 0
+	case sqlparser.SelectStatement:
+		return true
+	}
+	return false
+}
+
+func getColVindexes(insOp *Insert) (colVindexes []*vindexes.ColumnVindex) {
+	// For unsharded table the Column Vindex does not mean anything.
+	// And therefore should be ignored.
+	if !insOp.VTable.Keyspace.Sharded {
+		return
+	}
+	for _, colVindex := range insOp.VTable.ColumnVindexes {
+		if colVindex.IsPartialVindex() {
+			continue
+		}
+		colVindexes = append(colVindexes, colVindex)
+	}
+	return
+}
+
+func checkAndErrIfVindexChanging(setClauses sqlparser.UpdateExprs, col sqlparser.IdentifierCI) error {
+	for _, assignment := range setClauses {
+		if col.Equal(assignment.Name.Name) {
+			valueExpr, isValuesFuncExpr := assignment.Expr.(*sqlparser.ValuesFuncExpr)
+			// update on duplicate key is changing the vindex column, not supported.
+			if !isValuesFuncExpr || !valueExpr.Name.Name.Equal(assignment.Name.Name) {
+				return vterrors.VT12001("DML cannot update vindex column")
+			}
+			return nil
+		}
+	}
+	return nil
+}
+
+// findOrAddColumn finds the position of a column in the insert. If it's
+// absent it appends it to the with NULL values.
+// It returns the position of the column and also boolean representing whether it was added or already present.
+func findOrAddColumn(ins *sqlparser.Insert, col sqlparser.IdentifierCI) (int, bool) {
+	colNum := findColumn(ins, col)
+	if colNum >= 0 {
+		return colNum, false
+	}
+	colOffset := len(ins.Columns)
+	ins.Columns = append(ins.Columns, col)
+	if rows, ok := ins.Rows.(sqlparser.Values); ok {
+		for i := range rows {
+			rows[i] = append(rows[i], &sqlparser.NullVal{})
+		}
+	}
+	return colOffset, true
+}
+
+// findColumn returns the column index where it is placed on the insert column list.
+// Otherwise, return -1 when not found.
+func findColumn(ins *sqlparser.Insert, col sqlparser.IdentifierCI) int {
+	for i, column := range ins.Columns {
+		if col.Equal(column) {
+			return i
+		}
+	}
+	return -1
+}
+
+func populateInsertColumnlist(ins *sqlparser.Insert, table *vindexes.Table) *sqlparser.Insert {
+	cols := make(sqlparser.Columns, 0, len(table.Columns))
+	for _, c := range table.Columns {
+		cols = append(cols, c.Name)
+	}
+	ins.Columns = cols
+	return ins
+}
+
+// modifyForAutoinc modifies the AST and the plan to generate necessary autoinc values.
+// For row values cases, bind variable names are generated using baseName.
+func modifyForAutoinc(ins *sqlparser.Insert, vTable *vindexes.Table) (*Generate, error) {
+	if vTable.AutoIncrement == nil {
+		return nil, nil
+	}
+	gen := &Generate{
+		Keyspace:  vTable.AutoIncrement.Sequence.Keyspace,
+		TableName: sqlparser.TableName{Name: vTable.AutoIncrement.Sequence.Name},
+	}
+	colNum, newColAdded := findOrAddColumn(ins, vTable.AutoIncrement.Column)
+	switch rows := ins.Rows.(type) {
+	case sqlparser.SelectStatement:
+		gen.Offset = colNum
+		gen.added = newColAdded
+	case sqlparser.Values:
+		autoIncValues := make([]evalengine.Expr, 0, len(rows))
+		for rowNum, row := range rows {
+			// Support the DEFAULT keyword by treating it as null
+			if _, ok := row[colNum].(*sqlparser.Default); ok {
+				row[colNum] = &sqlparser.NullVal{}
+			}
+			expr, err := evalengine.Translate(row[colNum], nil)
+			if err != nil {
+				return nil, err
+			}
+			autoIncValues = append(autoIncValues, expr)
+			row[colNum] = sqlparser.NewArgument(engine.SeqVarName + strconv.Itoa(rowNum))
+		}
+		gen.Values = evalengine.NewTupleExpr(autoIncValues...)
+	}
+	return gen, nil
 }
