+++ go/vt/vtgate/planbuilder/operators/update.go
@@ -229,7 +229,7 @@ func getFKRequirementsForUpdate(ctx *plancontext.PlanningContext, updateExprs sq
 
 func buildFkOperator(ctx *plancontext.PlanningContext, updOp ops.Operator, updClone *sqlparser.Update, parentFks []vindexes.ParentFKInfo, childFks []vindexes.ChildFKInfo, updatedTable *vindexes.Table) (ops.Operator, error) {
 	// We only support simple expressions in update queries for foreign key handling.
-	if sqlparser.IsNonLiteral(updClone.Exprs) {
+	if updClone.Exprs.IsNonLiteral() {
 		return nil, vterrors.VT12001("update expression with non-literal values with foreign key constraints")
 	}
 
@@ -249,7 +249,7 @@ func splitChildFks(fks []vindexes.ChildFKInfo) (restrictChildFks, cascadeChildFk
 		// Any RESTRICT type foreign keys that arrive here for 2 reasons—
 		//	1. cross-shard/cross-keyspace RESTRICT cases.
 		//	2. shard-scoped/unsharded RESTRICT cases arising because we have to validate all the foreign keys on VTGate.
-		if isRestrict(fk.OnUpdate) {
+		if fk.OnUpdate.IsRestrict() {
 			// For RESTRICT foreign keys, we need to verify that there are no child rows corresponding to the rows being updated.
 			// This is done using a FkVerify Operator.
 			restrictChildFks = append(restrictChildFks, fk)
@@ -272,7 +272,7 @@ func createFKCascadeOp(ctx *plancontext.PlanningContext, parentOp ops.Operator, 
 
 	for _, fk := range childFks {
 		// We should have already filtered out update restrict foreign keys.
-		if isRestrict(fk.OnUpdate) {
+		if fk.OnUpdate.IsRestrict() {
 			return nil, vterrors.VT13001("ON UPDATE RESTRICT foreign keys should already be filtered")
 		}
